refactor NewEventForm into EventForm. EventForm is now a Formik-Yup form
