--- src/components/NewEventForm/index.tsx
@@ -1,3 +1,0 @@
-import NewEventForm from './NewEventForm';
-
-export default NewEventForm;
